Store colors as int16_t foreground and background instead of size_t prebuilt pairs
In ncurses "color pairs" are created and then used.  We used to register each face
as a separate color pair.  When doing this, we lost the ability to merge the colors
of multiple faces, which we want for future features.

Now we generate color pairs for all possible color combinations and then use colors
directly and recreate color pairs in the ncurses specific code.  The `Cell` stores
`foreground` and `background` colors directly.  When doing this we also changed
colors to be represented as `int16_t` with `-1` meaning default.  Most faces don't
specify the background color (setting it to `-1`), allowing for other faces to be
merged on top and modify the background color.