Completion_Cache::set_engine sets the state to INITIAL and returns when given the same engine again
This makes completion_engine's parameter is_initial_frame true when it
is invoked but has already loaded results.

This fixes a long standing bug that buffer_completion_engine wouldn't
reload the buffer list when the user launches buffer completion.
Example use case to trigger the bug:

Kill a buffer (loading completion) then switch buffers using the
existing completion state.  This would cause the killed buffer to look
as if it still was open.