Use Arc<Buffer_Handle> instead of Buffer_Id and move the id field
Change a bunch of references from Buffer_Id to Buffer_Handle.  This makes us
generally more safe against kill commands, which hasn't really been tested at all.
It also changes the behavior of each dialog that wants to manipulate the selected
buffer to assume that the buffer didn't change while the prompt was occurring.

But it does make all of these dialogs kill safe and makes goto_jump kill safe.

Maybe at some point in the future we should make some sort of save/restore layout functionality?

The id field is now in the Buffer instead of the Buffer_Handle.